Separate basis encode & transcode into separate files.
