--- lib/basisu_sgd.c
@@ -1,32 +1,0 @@
-/* -*- tab-width: 4; -*- */
-/* vi: set sw=2 ts=4 expandtab: */
-
-/*
- * ©2019 Khronos Group, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * @internal
- * @file basisu_sgd.c
- * @~English
- *
- * @brief Functions for creating Basis Universal supercompression global data.
- *
- * @author Mark Callow, www.edgewise-consulting.com
- */
-
-void basisuSgdCreate() {
-   // do something.
-}
