+++ lib/libktx.gypi
@@ -36,9 +36,9 @@
       'basisu/transcoder/basisu_file_headers.h',
       'basisu/transcoder/basisu_transcoder.cpp',
       'basisu/transcoder/basisu_transcoder.h',
-      'basisu_sgd.c',
-      'basisu_sgd.h',
-      'basisu.cpp',
+      'basis_sgd.h',
+      'basis_encode.cpp',
+      'basis_transcode.cpp',
       'checkheader.c',
       'dfdutils/createdfd.c',
       'dfdutils/dfd.h',
